shmsys: cast between incompatible function types
../../common/os/shm.c:188:2: error: cast between incompatible function types from 'uintptr_t (*)(int,  uintptr_t,  uintptr_t,  uintptr_t)' {aka 'long unsigned int (*)(int,  long unsigned int,  long unsigned int,  long unsigned int)'} to 'int (*)()' [-Werror=cast-function-type]
  (int (*)())shmsys
  ^
../../common/os/shm.c:195:2: error: cast between incompatible function types from 'uintptr_t (*)(int,  uintptr_t,  uintptr_t,  uintptr_t)' {aka 'long unsigned int (*)(int,  long unsigned int,  long unsigned int,  long unsigned int)'} to 'int (*)()' [-Werror=cast-function-type]
  (int (*)())shmsys
  ^
cc1: all warnings being treated as errors

+++ usr/src/uts/common/os/shm.c
@@ -185,14 +185,14 @@ static struct sysent ipcshm_sysent = {
 #else	/* _SYSCALL32_IMPL */
 	SE_ARGC | SE_NOUNLOAD | SE_32RVAL1,
 #endif	/* _SYSCALL32_IMPL */
-	(int (*)())shmsys
+	(int (*)())(uintptr_t)shmsys
 };
 
 #ifdef	_SYSCALL32_IMPL
 static struct sysent ipcshm_sysent32 = {
 	4,
 	SE_ARGC | SE_NOUNLOAD | SE_32RVAL1,
-	(int (*)())shmsys
+	(int (*)())(uintptr_t)shmsys
 };
 #endif	/* _SYSCALL32_IMPL */
